restart app
