+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "ignews",
+  "name": "ig.news",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -9,19 +9,17 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "next": "12.2.5",
+    "next": "13.2.4",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-icons": "^4.4.0",
-    "sass": "^1.54.5",
-    "stripe": "^10.7.0"
+    "react-icons": "^4.8.0",
+    "sass": "^1.60.0"
   },
   "devDependencies": {
-    "@types/next": "^9.0.0",
-    "@types/node": "^18.7.13",
-    "@types/react": "^18.0.17",
-    "eslint": "8.22.0",
-    "eslint-config-next": "12.2.5",
-    "typescript": "^4.7.4"
+    "@types/node": "^18.15.11",
+    "@types/react": "^18.0.31",
+    "eslint": "8.36.0",
+    "eslint-config-next": "13.2.4",
+    "typescript": "^5.0.3"
   }
 }
