+++ src/components/Header/styles.module.scss
@@ -3,7 +3,6 @@
   border-bottom: 1px solid var(--gray-800);
 }
 
-
 .headerContent {
   max-width: 1120px;
   height: 5rem;
@@ -27,7 +26,7 @@
 
       transition: color 0.2s;
 
-      & + a {
+      &+a {
         margin-left: 2rem;
       }
