--- src/components/SignInButton/index.tsx
@@ -1,27 +1,0 @@
-import { FaGithub } from 'react-icons/fa'
-import { FiX } from 'react-icons/fi'
-
-import styles from './styles.module.scss';
-
-export function SignInButton() {
-  const isUserLoggedIn = true;
-
-  return isUserLoggedIn ? (
-    <button
-      type="button"
-      className={styles.signInButton}
-    >
-      <FaGithub color='#04d361' />
-      Tiago Mota
-      <FiX color='#737380' className={styles.closeIcon} />
-    </button>
-  ) : (
-    <button
-      type="button"
-      className={styles.signInButton}
-    >
-      <FaGithub color='#eba417' />
-      Sign in width Github
-    </button>
-  );
-}
