+++ src/components/SubscribeButton/index.tsx
@@ -1,17 +1,14 @@
-import React from 'react';
+import React from "react";
 
-import styles from './styles.module.scss';
+import styles from "./styles.module.scss";
 
-interface SubscribeButtonProps {
-  priceId: string;
-}
+// interface SubscribeButtonProps {
+//   priceId: string;
+// }
 
-export function SubscribeButton({ priceId }: SubscribeButtonProps) {
+export function SubscribeButton() {
   return (
-    <button
-      type="button"
-      className={styles.subscribeButton}
-    >
+    <button type="button" className={styles.subscribeButton}>
       Subscribe now
     </button>
   );
