+++ src/pages/_app.tsx
@@ -1,7 +1,6 @@
-import { AppProps } from 'next/app';
-import Header from '../components/Header/inde';
-
-import "../styles/global.scss";
+import { AppProps } from "next/app";
+import "../../styles/global.scss";
+import Header from "../components/Header/Header";
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
@@ -9,7 +8,7 @@ function MyApp({ Component, pageProps }: AppProps) {
       <Header />
       <Component {...pageProps} />
     </>
-  )
+  );
 }
 
-export default MyApp
+export default MyApp;
