+++ src/pages/home.module.scss
@@ -40,6 +40,6 @@
   }
 
   button {
-    margin-top: 2.5rem;
+    margin-top: 2rem;
   }
 }
